changed

+++ Codeforces/1244/A/A.cpp
@@ -29,7 +29,7 @@ int main() {
   T = read();
   while (T--) {
     a = read(), b = read(), c = read(), d = read(), k = read();
-    int x = ceil(a / c), y = ceil(b / d);
+    int x = ceil((double)a / c), y = ceil((double)b / d);
     if (x + y > k)
       puts("-1");
     else
